move validation data type classes on japan module
