+++ lib/zakuro/era/japan/version/resource/validator.rb
@@ -2,6 +2,10 @@
 
 require_relative '../../../japan/calendar'
 require_relative '../../../western/calendar'
+
+require_relative '../../type/validation/both/date'
+require_relative '../../type/validation/both/year'
+
 require_relative './type'
 require 'yaml'
 
@@ -64,7 +68,7 @@ module Zakuro
             # @return [Array<String>] 不正メッセージ
             #
             def validate_last_year
-              Both::Year.new(hash: last_year).validate
+              Type::Validation::Both::Year.new(hash: last_year).validate
             end
 
             #
@@ -73,7 +77,7 @@ module Zakuro
             # @return [Array<String>] 不正メッセージ
             #
             def validate_last_date
-              Both::Date.new(hash: last_date).validate
+              Type::Validation::Both::Date.new(hash: last_date).validate
             end
 
             #
@@ -172,7 +176,7 @@ module Zakuro
             # @return [Array<String>] 不正メッセージ
             #
             def validate_start_year
-              Both::Year.new(hash: start_year).validate
+              Type::Validation::Both::Year.new(hash: start_year).validate
             end
 
             #
@@ -181,7 +185,7 @@ module Zakuro
             # @return [Array<String>] 不正メッセージ
             #
             def validate_start_date
-              Both::Date.new(hash: start_date).validate
+              Type::Validation::Both::Date.new(hash: start_date).validate
             end
 
             #
@@ -194,149 +198,6 @@ module Zakuro
               return false if released.nil?
 
               released.is_a?(TrueClass) || released.is_a?(FalseClass)
-            end
-          end
-
-          #
-          # Both 和暦/西暦
-          #
-          module Both
-            #
-            # Year 年
-            #
-            class Year
-              # @return [String] 和暦元号年
-              attr_reader :japan
-              # @return [String] 西暦年
-              attr_reader :western
-
-              #
-              # 初期化
-              #
-              # @param [Hash<String, Strin>] hash 年情報
-              #
-              def initialize(hash:)
-                @japan = hash['japan']
-                @western = hash['western']
-              end
-
-              #
-              # 検証する
-              #
-              # @return [Array<String>] 不正メッセージ
-              #
-              def validate
-                failed = []
-
-                failed.push("invalid japan year. #{japan}") unless japan?
-
-                failed.push("invalid western year. #{western}") unless western?
-
-                failed
-              end
-
-              #
-              # 和暦元号年を検証する
-              #
-              # @return [True] 正しい
-              # @return [False] 正しくない
-              #
-              def japan?
-                return false unless @japan
-
-                japan.is_a?(Integer)
-              end
-
-              #
-              # 和暦元号年を検証する
-              #
-              # @return [True] 正しい
-              # @return [False] 正しくない
-              #
-              def western?
-                return false unless @western
-
-                western.is_a?(Integer)
-              end
-            end
-
-            #
-            # Date 日
-            #
-            class Date
-              # @return [String] 和暦日
-              attr_reader :japan
-              # @return [String] 西暦日
-              attr_reader :western
-              # @return [True] 省略可
-              # @return [False] 省略不可
-              attr_reader :optional
-
-              #
-              # 初期化
-              #
-              # @param [Hash<String, Strin>] hash 日情報
-              #
-              def initialize(hash:, optional: false)
-                @japan = hash['japan']
-                @western = hash['western']
-                @optional = optional
-              end
-
-              #
-              # 検証する
-              #
-              # @return [Array<String>] 不正メッセージ
-              #
-              def validate
-                failed = []
-
-                failed.push("invalid japan date. #{japan}") unless japan?
-
-                failed.push("invalid western date. #{western}") unless western?
-
-                failed
-              end
-
-              #
-              # 和暦日を検証する
-              #
-              # @return [True] 正しい
-              # @return [False] 正しくない
-              #
-              def japan?
-                return true if optional?(text: japan)
-
-                Japan::Calendar.valid_date_text(text: japan)
-              end
-
-              #
-              # 西暦日を検証する
-              #
-              # @return [True] 正しい
-              # @return [False] 正しくない
-              #
-              def western?
-                return true if optional?(text: western)
-
-                Western::Calendar.valid_date_text(text: western)
-              end
-
-              #
-              # 省略可で省略されているか
-              #
-              # @param [String] text 文字列
-              #
-              # @return [True] 省略あり
-              # @return [False] 省略なし
-              #
-              def optional?(text: '')
-                return false unless optional
-
-                return true if text == ''
-
-                false
-              end
             end
           end
